+++ CHANGELOG.rst
@@ -3,6 +3,7 @@
 
     * ``SignalReader`` base class and ``CsvReader``.
     * ``Signal.plot``: grid on by default. New parameter ``t0``.
+    * ``to_csv``: simple CSV export
 
     Fixes: